QoL: Push cursor to the end of the argument when changing #char argument.
For instance, first using CTRL+A to set the character number after
a #char comand to 99, then using CTRL+A again to set it to 100 would
leave the cursor at the last digit, which would cause it to be cut off
if one were to press enter. This change always places it after the last
digit so that pressing enter minimizes one's surprise.